refactor : mapping method 위치 변경 -> 엔티티가 DTO에 의존하지 않게 하였음
