+++ src/main/java/kr/co/springbootweeklymission/member/api/dto/response/MemberResDTO.java
@@ -1,5 +1,6 @@
 package kr.co.springbootweeklymission.member.api.dto.response;
 
+import kr.co.springbootweeklymission.member.domain.entity.Member;
 import kr.co.springbootweeklymission.member.domain.model.MemberStatus;
 import lombok.*;
 
@@ -14,5 +15,12 @@ public class MemberResDTO {
     public static class READ {
         private UUID memberId;
         private MemberStatus memberStatus;
+
+        public static MemberResDTO.READ toMemberReadDto(Member member) {
+            return MemberResDTO.READ.builder()
+                    .memberId(member.getMemberId())
+                    .memberStatus(member.getMemberStatus())
+                    .build();
+        }
     }
 }
