+++ src/main/java/kr/co/springbootweeklymission/member/domain/entity/Member.java
@@ -1,14 +1,11 @@
 package kr.co.springbootweeklymission.member.domain.entity;
 
-import kr.co.springbootweeklymission.member.api.dto.response.MemberResDTO;
 import kr.co.springbootweeklymission.member.domain.model.MemberStatus;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.util.UUID;
 
+@Getter
 @EqualsAndHashCode
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class Member {
@@ -20,13 +17,6 @@ public class Member {
                    MemberStatus memberStatus) {
         this.memberId = memberId;
         this.memberStatus = memberStatus;
-    }
-
-    public static MemberResDTO.READ toMemberReadDto(Member member) {
-        return MemberResDTO.READ.builder()
-                .memberId(member.memberId)
-                .memberStatus(member.memberStatus)
-                .build();
     }
 
     public boolean isBlackMember() {
