+++ src/main/resources/files/voucher_file.txt
@@ -1,3 +1,8 @@
 988b4b0f-1f87-4a6a-96a7-8187f035ad83 0 FIXED_DISCOUNT
 7e3bc446-64f4-4ae0-96d5-f92f5bd34644 0 FIXED_DISCOUNT
 b856f3b0-8d25-4727-ab1a-57a71b23c867 0 FIXED_DISCOUNT
+c3fbf5f4-6bde-4d32-89a4-a609985afae0 100 FIXED_DISCOUNT
+adc004a5-8034-429d-9ad8-c137f64117ec 100 FIXED_DISCOUNT
+a28fc164-3716-462e-9523-0dd9b584cab9 0 FIXED_DISCOUNT
+882ce497-f84c-49b1-a8b6-3cbf9a7eead9 0 FIXED_DISCOUNT
+42e2f1f6-9b5c-4ee1-8267-a078de6930a7 0 FIXED_DISCOUNT
